requirements.txt
EbookLib-MD
beautifulsoup4
prettytable
openai
python-dotenv
pyttsx3
